renamed variables
Usually the word cooldown is used for effects that repeat, for example the duration between burn damage ticks is indeed a cooldown.

But the time an effect lasts is usually referred to as a duration. So the time the burn sticks is the burn duration.

Renaming the variables such that durations are called durations rather than cooldowns should make the code easier to read and maintain for others.