bug fix: poison gas applying 2 stacks of poison if target is at 0 stacks.
With the old order of operrations poison gas would trigger on the target if it has 0 stacks. This alters the state of the pokemon, giving it 1 stack of poison meaning the updatePoison method triggers, which in turn will activate poison gas again.

Moving the condition to go after updatePoison should fix this issue.