removed redundancy in update methods
The only reason why its necessary to check duration - dt <= 0 is because the dt is only subtracted otherwise in the else statement.

By moving the duration - dt application out of the else statement it becomes possible to check just duration <= 0, making it easier to read what is happening and ensures that the duration is actually reduced to or below 0 for consistency purposes.