Fix Desc and % damage
Fixed a discrepency between the descriptions and the ability damage. It's likely that these abilities will further change as Dasocks brings up a good point about potential damage output:
https://discord.com/channels/737230355039387749/1300650743396827169/1423703144411430942

That being said, I'm concerned about correctness not balance. If further changes are to be made, that is the dev's perogative.