Disable duplicate pokemon from being selected
John Rei asked me to look in to this and I think I found the issue. Added a check to ensure duplicate pokemon cannot be selected in carousels (primarily on t0) in which previous selections can make this possible.
⁠unknown

The bug here works as follows:
On turn 0 if you enter a steel/arti/rock portal, those synergies will be rolled for candidates. If steel rolls klinkz, then arti will be out of candidates to roll. It is not guaranteed to roll kilnkz again but there is no check against it. As a result, you have a chance to see duplicate pokemon. This can happen in any scenario where a previous roll prevents a subsequent roll from having candidates. This can happen if a ground symbol rolls swinub, followed by an ice symbol that no longer has candidates. It can also happen with any grass flora portal if grass rolls sprigatitto first. It can also happen with any portal that has a fossil symbol because they have no candidates by default.

This should not be a problem on t10 or t20, as every pool has at least 3 pokemon in it. Even with four portal symbols the only problem I forsee is if a dragon, monster, and flying symbol (or any combination of those with ground) with a ground fourth symbol rolls every ground legendary leaving no candidates for the final (ground) symbol. This seems such an unlikely occurrence though that I am not surprised no one has ever reported experiencing it.

The intended fix is if candidates.length === 0, allow candidates to be all pokemon regardless of synergy while still discarding pokemon not in their region or already selected.

It's possible I am misunderstanding parts of the selection process in which case this can be closed, but I suspect this is the cause of the bug and I hope this PR can fix it.